Add "valid" rule that checks if the SVG is valid XML (#6)
* Add "valid" rule that checks if the SVG is valid XML

In light of [1] this adds XML validation to the linter. Since none of
the alternative parsers to htmlparser2 (listed in [2]) seem to fail on
the example given in [1], I decided to use fast-xml-parser ([3]) for
this. I implemented as a "rule" that is always enabled through the
default configuration.

[1] https://github.com/simple-icons/simple-icons/pull/1112
[2] https://www.npmjs.com/package/htmlparser2#performance
[3] https://www.npmjs.com/package/fast-xml-parser

* Minor cleanup in valid rule

Use the correct name for the logger instance
Remove leftover console.log

* Update Valid rule

... according to @birjolaxew's comments

* Improve Valid rule test suite

* Slightly change tests for 'valid'

Previously checked default by not giving a config
Now checks default by giving an empty rules object
This ensures it also tests merging of rule object with defaults

* Support merging of default rules and passed rules object

Previously if any rules were set, all default rules would be ignored
Now merges rules object such that defaults are applied (but can be
overwritten)

* Make 'valid' succeed on empty SVG