Add .npmignore
Add an .npmignore file to ignore certain files/folders when publishing.
The .npmignore uses a whitelist style, first ignoring everything and
then listing only the things that should be published. Most notably this
will remove dot-files and folders as well as the tests.

I kept example.png because it's used in the README. It can be removed
later if that's preferred.