feat: exit code 1 when specified config file is missing
Update the CLI such that *if* a configuration file path is specified,
i.e. the `--config` option is used, then the configuration file must
exist. If it does not exist, the CLI will exist with a non-zero exit
code, after logging an error.

If `--config` is not used the CLI behaviour is unchanged.