feat: make configuration optional (#40)
* test: run SVGLint API without config

Add a test for each of the SVGLint API functions to verify they work if
no config is provided - with the expected behaviour being the same as
that of providing the config `{}` explicitly.

* feat: update API to make config optional

Update the API so that the linting configuration is optional. Currently,
the config falls back to an empty object.

* feat: don't require config file with CLI

Update the SVGLint CLI to not require a configuration file. The
`getConfigurationFile` function was updated to output a special value
(`false`) if no configuration file is found. When the CLI, using this
function, detects there's no config file it will not try to read
(`require`) that file.

* chore(docs): add note about the default configuration

* feat: exit code 1 when specified config file is missing

Update the CLI such that *if* a configuration file path is specified,
i.e. the `--config` option is used, then the configuration file must
exist. If it does not exist, the CLI will exist with a non-zero exit
code, after logging an error.

If `--config` is not used the CLI behaviour is unchanged.