refactor: update source to ES modules
Update the source code from CommonJS to ES modules style. This consists
of changing `require`s to `import from` statements, `module.exports` to
`export` statements, and in some cases the structure of imports and
exports as well. Notably when the old `require` statement also contained
some additional logic, which is not possible with import statements.