feat: add CJS version of Node API
Add a CommonJS version of the Node API of SVGLint by building it using
Rollup [1]. Update the package manifest to allow for importing/requiring
ESM and CJS (resp.) and `.gitignore` to ignore the generated CJS file.

One "custom" Rollup plugin (provided by Rollup but not included by
default) is used, namely `plugin-dynamic-import-vars` [2]. This plugin
ensures that dynamic imports are properly handled as the CJS output
generated by Rollup cannot dynamically import ESM rules. To make this
work the rule-loader had to be updated as well, unfortunately the
requirements of this plugin are quite strict and it is required that the
`"../rules"` path is hardcoded into the `import` statement. Furthermore
it has trouble detecting things that should be ignored, so errors are
disabled to prevent errors due to dynamic imports of external plugins.

--
1. https://rollupjs.org/guide/en/
2. https://www.npmjs.com/package/@rollup/plugin-dynamic-import-vars