test: run CLI on valid and invalid SVG
Add a test case for running the CLI on a valid SVG and a test case for
running the CLI on an invalid SVG.

Update the `execCliWith` helper function to return the error if an error
is caught. This follows the documentation of execa [1], the error object
has the same structure as the resolved value of the `execa` function.

--
1. https://github.com/sindresorhus/execa/tree/e4b929547ade335b0dd9d7afe1a35989e19fbc8a#handling-errors