ci: update action versions and configs, and token perms (#48)
* ci: update actions/setup-node from v1 to v2

And enable the cache functionality built into v2.

* ci: use `ubuntu-latest` instead of `ubuntu-18.04`

* ci: remove `fetch-depth: 0` where it's unnecessary

The `fetch-depth: 0` option of actions/checkout results in the entire
repository being pulled down. This is often unnecessary, so it is
removed for the testing workflow. For now it's kept in the release
workflow as semantic-release seems to require the full history.

* ci: configure permissions for GITHUB_TOKEN

Add in-workflow configuration of the permissions of the GITHUB_TOKEN.
The default is set to `read-all` as this project is open source anyway,
the release workflow requires write access to contents, issues, and pulls,
to be able to create a GitHub release, comment on released issues, and
comment on released pull requests (resp.).

You can read more about permissions here:
https://docs.github.com/en/actions/learn-github-actions/workflow-syntax-for-github-actions#permissions