Add test suite for the CLI (#49)
* chore(deps): add execa to test the CLI in a test suite

* test: proof of concept test suite for the CLI

* test: run CLI on valid and invalid SVG

Add a test case for running the CLI on a valid SVG and a test case for
running the CLI on an invalid SVG.

Update the `execCliWith` helper function to return the error if an error
is caught. This follows the documentation of execa [1], the error object
has the same structure as the resolved value of the `execa` function.

* ci: remove test job to run the CLI

The `test-cli` job is redundant now that there's a test suite that runs
the CLI.
The `test` job now also runs once on macOS and once on Windows for
node-version 16 (matching the removed `test-cli` job).

* ci: lint source code only once

Lint the source code only once instead of on every OS-NodeJS version
combination. This both improves the running time of the CI as well as
avoids problems the CI is having when running the linter on
Windows (incorrect line endings).

* ci: remove manual invocation of the CLI in the test job

--
1. https://github.com/sindresorhus/execa/tree/e4b929547ade335b0dd9d7afe1a35989e19fbc8a#handling-errors