fix(deps): bump strip-ansi and log-update (#51)
* chore(deps): upgrade strip-ansi

Major version bumps are all for increases in the minimum required NodeJS
version. The minimum required NodeJS version for v7 is below that of
this package, hence it's safe to upgrade.

* chore(deps): upgrade log-update

Major version bumps increase in the minimum required NodeJS version. The
minimum required NodeJS version for v5 is below that of
this package, hence it's safe to upgrade.

Additionally, the v5 major release made some changes to the API, but
this does not affect how this package uses log-update.