refactor: use `fs.access` over `fs.exists` (#61)
The `fs.exists` function is deprecated. `fs.access` is one of the
recommended replacements - the other being `fs.stat`, which seems to be
doing too much for what is needed here.

To keep behavior as similar to `fs.exists` as possible, we only check
for the `F_OK` property (is the file visible to the process) rather than
e.g. the `R_OK` property (is the file readable).