ci: bump lowest Node.js version tested
While 12.20.0 and 14.13.1 are still supported by this package, some
dependencies minimum required version is 12.22.0 and 14.17.0 (resp.).

I don't expect these couple of minor bumps in Node.js version to cause a
compatibility issue in this project, so as a quick solution I'm bumping
the tested versions to match the lowest supported Node.js versions by
dependencies.