Test behaviour when no configuration file is present
Update the CLI tests to include a test for the case where a project does
not have a configuration file. This test currently fails because this is
currently not supported due to a bug introduced in v2.2.0.

To accomodate testing this, the configuration file that was present at
the root of the project has been moved and related tests updated
accordingly. This is necessary because svglint automatically searches
for a configuration file up the file tree, so having a configuration
file at the root of the project makes it impossible to test the scenario

The new tests matches against a string because unfortunately the error
is silent - it doesn't exit with a non-zero exit code. Improvements to
this are welcome as contributions.