fix: support no configuration file for the CLI
Update the CLI to work if no configuration file is found.

This is the simplest fix I could think of. Theoretically the problem
could be addressed in the API (svglint.js line 137 and 150) by replacing
the default parameter `config={}` with an implementation that defaults
to an empty object for both `undefined` **and** `null`. However, it's
currently not clear that's the right move.

Another alternative would be for `loadConfigurationFile` to return
`undefined` instead of `null`. However, that could interfere with the
intended behaviour of that function, making it impossible to distinguish
a missing default export from no-configuration-file-found.