fix: support running CLI in project without config file (#80)
* Test behavior when no configuration file is present

Update the CLI tests to include a test for the case where a project does
not have a configuration file. This test currently fails because this is
currently not supported due to a bug introduced in v2.2.0.

To accommodate testing this, the configuration file that was present at
the root of the project has been moved and related tests updated
accordingly. This is necessary because svglint automatically searches
for a configuration file up the file tree, so having a configuration
file at the root of the project makes it impossible to test the scenario

The new tests matches against a string because unfortunately the error
is silent - it doesn't exit with a non-zero exit code. Improvements to
this are welcome as contributions.

* Support no configuration file for the CLI

Update the CLI to work if no configuration file is found.

This is the simplest fix I could think of. Theoretically the problem
could be addressed in the API (svglint.js line 137 and 150) by replacing
the default parameter `config={}` with an implementation that defaults
to an empty object for both `undefined` **and** `null`. However, it's
currently not clear that's the right move.

Another alternative would be for `loadConfigurationFile` to return
`undefined` instead of `null`. However, that could interfere with the
intended behavior of that function, making it impossible to distinguish
a missing default export from no-configuration-file-found.