fix: address high severity issue by upgrading fast-xml-parser
Upgrade fast-xml-parser from v3 to v4 in order to address a (according
to npm) high severity vulnerability, namely CVE-2023-34104. I did not
evaluate the impact of the CVE on this project, but upgrading is easy
enough and avoids any potential problems. Plus, staying on the latest
available version is always good :)