feat: lint SVG from stdin (#82)
Extend the CLI to support providing an SVG on `stdin` [1]. The CLI will
only do this when provided with the `--stdin` flag, I choose this design
as it was the only reliable way of switching between files and stdin I
could find.

Despite the awkward diff, the CLI flow for files (i.e. if `--stdin` is
not used) is unchanged. The stdin flow is based on [2] and just calls
the JS API's `lintSource` function on the entire input.

--
1. https://nodejs.org/api/process.html#processstdin
2. https://nodejs.org/api/stream.html#readablereadsize