ci: always use Node.js v20
In addition to testing on Node.js v20, use it for all CI jobs.

For the "Test / Lint" job, I tested locally if linting on Node.js v20
works and it does. For "Release / Semantic release" I verified that
`semantic-release` release supports Node.js v20 and as of v21.0.6 it
does.